REFA-0001 - refactor core data package file structure
