+++ core/src/main/java/ru/easycode/words504/data/cache/preferences/ProvideSharedPreferences.kt
@@ -1,4 +1,4 @@
-package ru.easycode.words504.data.providers.preferences
+package ru.easycode.words504.data.cache.preferences
 
 import android.content.Context
 import android.content.SharedPreferences
