+++ core/src/main/java/ru/easycode/words504/data/cache/storage/Storage.kt
@@ -1,4 +1,4 @@
-package ru.easycode.words504.data
+package ru.easycode.words504.data.cache.storage
 
 interface Storage : SimpleStorage, ObjectStorage {
