+++ core/src/main/java/ru/easycode/words504/data/cache/storage/StringStorage.kt
@@ -1,4 +1,4 @@
-package ru.easycode.words504.data
+package ru.easycode.words504.data.cache.storage
 
 interface StringStorage {
     fun read(key: String, default: String): String
