+++ core/src/main/java/ru/easycode/words504/data/cloud/retrofit/ProvideRetrofitBuilder.kt
@@ -1,4 +1,4 @@
-package ru.easycode.words504.data.providers.retrofit
+package ru.easycode.words504.data.cloud.retrofit
 
 import retrofit2.Retrofit
